% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_mean.R
\name{pw_mean}
\alias{pw_mean}
\title{Element-wise mean of two vectors a and b}
\usage{
pw_mean(a,b, na.rm = TRUE)
}
\arguments{
\item{a, b}{two vectors}

\item{na.rm}{logical indicating whether \code{NA} values should be ignored when computing means}
}
\description{
Element-wise mean of two vectors a and b (https://stackoverflow.com/questions/13990125/sample-from-vector-of-varying-length-including-1)
}
