% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_vect.R
\name{rand_vect}
\alias{rand_vect}
\title{Random vector of N integers that sum to M}
\usage{
rand_vect(N,M)
}
\arguments{
\item{N}{Size of the vector.}

\item{M}{Sum of elements in the vector.}

\item{sd}{Approximate standard deviation for elements of vector (default = 1).}

\item{pos.only}{Resulting vector contains only positive integers (default = TRUE)}
}
\value{
Vector of size \code{N} with elements summing to \code{M}
}
\description{
Random vector of N integers that sum to M (https://stackoverflow.com/questions/24845909/generate-n-random-integers-that-sum-to-m-in-r)
}
\examples{
rand_vect(3, 50)
[1] 17 16 17
rand_vect(10, 10, pos.only = FALSE)
[1]  0  2  3  2  0  0 -1  2  1  1
rand_vect(10, 5, pos.only = TRUE)
[1] 0 0 0 0 2 0 0 1 2 0
rand_vect(10, -1, pos.only = FALSE)
[1] -1 -1  1 -2  2  1  1  0 -1 -1
}
