% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.vec.R
\name{sample.vec}
\alias{sample.vec}
\title{Samples n elements from vector x even if x is size 1 (https://stackoverflow.com/questions/13990125/sample-from-vector-of-varying-length-including-1)}
\usage{
sample.vec(x,n)
}
\arguments{
\item{x}{vector to sample from}

\item{size}{a non-negative integer giving the number of items to choose.}

\item{replace}{should sampling be with replacement? (default=FALSE)}

\item{prob}{a vector of probability weights for obtaining the elements of the vector being sampled.}

\item{useHash}{logical indicating if the hash-version of the algorithm should be used. Can only be used for replace = FALSE, prob = NULL, and size <= n/2, and really should be used for large n, as useHash=FALSE will use memory proportional to n.}
}
\description{
Samples n elements from vector x even if x is size 1 (https://stackoverflow.com/questions/13990125/sample-from-vector-of-varying-length-including-1)
}
